+++ src/main/kotlin/com/atlassian/performance/tools/report/api/judge/RelativeNonparametricPerformanceJudge.kt
@@ -4,7 +4,6 @@ import com.atlassian.performance.tools.jiraactions.api.ActionType
 import com.atlassian.performance.tools.report.ActionMetricsReader
 import com.atlassian.performance.tools.report.api.ShiftedDistributionRegressionTest
 import com.atlassian.performance.tools.report.api.junit.FailedAssertionJUnitReport
-import com.atlassian.performance.tools.report.api.junit.JUnitReport
 import com.atlassian.performance.tools.report.api.junit.SuccessfulJUnitReport
 import com.atlassian.performance.tools.report.api.result.EdibleResult
 
@@ -19,8 +18,7 @@ class RelativeNonparametricPerformanceJudge(
         baselineResult: EdibleResult,
         experimentResult: EdibleResult
     ): Verdict {
-        val testReports = mutableListOf<JUnitReport>()
-        val failedActions = mutableListOf<ActionType<*>>()
+        val actionReports = mutableListOf<ActionReport>()
         for ((action, toleranceRatio) in toleranceRatios) {
             val actionReport = judge(
                 action,
@@ -28,12 +26,9 @@ class RelativeNonparametricPerformanceJudge(
                 baselineResult,
                 experimentResult
             )
-            testReports.add(actionReport.report)
-            if (actionReport.nonExceptionalFailure) {
-                failedActions.add(actionReport.action)
-            }
+            actionReports.add(actionReport)
         }
-        return Verdict(testReports, failedActions)
+        return Verdict(actionReports)
     }
 
     private fun judge(
@@ -41,43 +36,40 @@ class RelativeNonparametricPerformanceJudge(
         toleranceRatio: Float,
         baselineResult: EdibleResult,
         experimentResult: EdibleResult
-    ): ActionReport {
+    ): ActionReportImpl {
         val label = action.label
         val baselineCohort = baselineResult.cohort
         val experimentCohort = experimentResult.cohort
         val reportName = "Regression for $label $experimentCohort vs $baselineCohort"
         val reader = ActionMetricsReader()
         val baseline = reader.read(baselineResult.actionMetrics)[label]?.stats?.values
-            ?: return ActionReport(
+            ?: return ActionReportImpl(
                 report = FailedAssertionJUnitReport(reportName, "No action $label results for $baselineCohort"),
-                action = action
+                action = action,
+                nonExceptional = false
             )
         val experiment = reader.read(experimentResult.actionMetrics)[label]?.stats?.values
-            ?: return ActionReport(
+            ?: return ActionReportImpl(
                 report = FailedAssertionJUnitReport(reportName, "No action $label results for $experimentCohort"),
-                action = action
+                action = action,
+                nonExceptional = false
             )
         val test = ShiftedDistributionRegressionTest(baseline, experiment, mwAlpha = significance, ksAlpha = 0.0)
         return if (test.isExperimentRegressed(toleranceRatio.toDouble())) {
             val message = "Regression in [$label] is larger than allowed ${toleranceRatio.toPercentage()} tolerance at $significance significance level"
-            ActionReport(
+            ActionReportImpl(
                 report = FailedAssertionJUnitReport(reportName, message),
                 action = action,
-                nonExceptionalFailure = true
+                nonExceptional = true
             )
         } else {
-            ActionReport(
+            ActionReportImpl(
                 report = SuccessfulJUnitReport(reportName),
-                action = action
+                action = action,
+                nonExceptional = false
             )
         }
     }
 
     private fun Float.toPercentage(): String = "%+.2f%%".format(this * 100)
 }
-
-internal data class ActionReport(
-    val report: JUnitReport,
-    val action: ActionType<*>,
-    val nonExceptionalFailure: Boolean = false
-)