+++ src/main/kotlin/com/atlassian/performance/tools/report/api/judge/RelativeNonparametricStabilityJudge.kt
@@ -4,7 +4,6 @@ import com.atlassian.performance.tools.jiraactions.api.ActionType
 import com.atlassian.performance.tools.report.ActionMetricsReader
 import com.atlassian.performance.tools.report.api.ShiftedDistributionRegressionTest
 import com.atlassian.performance.tools.report.api.junit.FailedAssertionJUnitReport
-import com.atlassian.performance.tools.report.api.junit.JUnitReport
 import com.atlassian.performance.tools.report.api.junit.SuccessfulJUnitReport
 import com.atlassian.performance.tools.report.api.result.EdibleResult
 
@@ -19,58 +18,54 @@ class RelativeNonparametricStabilityJudge(
         baselineResult: EdibleResult,
         experimentResult: EdibleResult
     ): Verdict {
-        val testReports = mutableListOf<JUnitReport>()
-        val failedActions = mutableListOf<ActionType<*>>()
+        val actionReports = mutableListOf<ActionReport>()
         actions.forEach { action ->
             val actionReport = judge(
                 action,
                 baselineResult,
                 experimentResult
             )
-            testReports.add(actionReport.report)
-            if (actionReport.nonExceptionalFailure) {
-                failedActions.add(actionReport.action)
-            }
+            actionReports.add(actionReport)
         }
 
-        return Verdict(
-            reports = testReports,
-            failedActions = failedActions
-        )
+        return Verdict(actionReports)
     }
 
     private fun judge(
         action: ActionType<*>,
         baselineResult: EdibleResult,
         experimentResult: EdibleResult
-    ): ActionReport {
+    ): ActionReportImpl {
         val label = action.label
         val baselineCohort = baselineResult.cohort
         val experimentCohort = experimentResult.cohort
         val reportName = "Stability regression for $label $experimentCohort vs $baselineCohort"
         val reader = ActionMetricsReader()
         val baseline = reader.read(baselineResult.actionMetrics)[label]?.stats?.values
-            ?: return ActionReport(
+            ?: return ActionReportImpl(
                 report = FailedAssertionJUnitReport(reportName, "No action $label results for $baselineCohort"),
-                action = action
+                action = action,
+                nonExceptional = false
             )
         val experiment = reader.read(experimentResult.actionMetrics)[label]?.stats?.values
-            ?: return ActionReport(
+            ?: return ActionReportImpl(
                 report = FailedAssertionJUnitReport(reportName, "No action $label results for $experimentCohort"),
-                action = action
+                action = action,
+                nonExceptional = false
             )
         val test = ShiftedDistributionRegressionTest(baseline, experiment, mwAlpha = 0.0, ksAlpha = significance)
         return if (!test.equalDistributionsAfterShift) {
             val message = "[$label] distribution shapes are different at $significance significance level"
-            ActionReport(
+            ActionReportImpl(
                 report = FailedAssertionJUnitReport(reportName, message),
                 action = action,
-                nonExceptionalFailure = true
+                nonExceptional = true
             )
         } else {
-            ActionReport(
+            ActionReportImpl(
                 report = SuccessfulJUnitReport(testName = reportName),
-                action = action
+                action = action,
+                nonExceptional = false
             )
         }
     }